New override Connect() method.
It recently emerged on our platform the necessity to disable the Clean Session functionality, but unfortunately there is no override method for Connect() which would only take in clientId and cleanSession as arguments. From our testing, the connection attempt fails when passing null username and password, so we can't use any of the existing override methods. The reason why the Connect() method can succefully connect when passing the clientId as the only argument (hence username and password will be null) is still unclear.